Assert the state of all repositories in the storage
TransactionManager's tests currently assert only the state of a single
repository in the storage. As we are about to add support for multiple
repositories in partition, we should assert that all repositories in the
storage match the expected state. This commit does so by wiring the new
test helper to do so in the existing tests. While we still have only a
single repository in each test, we are now asserting the fact that there's
only that single repository, and support asserting multiple repositories.