Begin transaction against a given repository
This commit extends the Begin method to take in a relative path and
start a transaction against that repository. With this change it becomes
possible to operate on multiple repositories within given partition.
As it is now possible to specify which repository the transaction targets,
the relativePath field of the TransactionManager is removed.