Allow modifying storage prior to TransactionManager running
TransactionManager's tests have a helper to modify a repository
prior on the disk manually prior to the manager starting. This used
to be sufficient when the manager only supported a single repository.
Now that the manager supports multiple repositories, let's expand the
scope of this function to allow modifying the storage. This way we can
better test functionality that spans multiple repositories.