Snapshot repository's alternate as well
Some repositories may have an 'objects/info/alternates' file containing
a relative path from the repository's object directory to another
repository's object directory. Git reads the file to find the other
repository's object database and reads objects also from there. This
is used in Gitaly to implement object pooling where objects are shared
and deduplicated between multiple repositories.

TransactionManager isn't currently considering this. Snapshotting a
repository without its object pool may end up snapshotting the repository
in a corrupt state as some of the objects the repository needs were in
the alternate and not available in the snapshot. This commit add snapshotting
support for pooled repositories. After snapshotting the transaction's
target repository, the snapshotted repository's alternates file is read.
If it contains a path, the pointed to repository is also included in the
transaction's snapshot. As the repositories are stored in the snapshot
at their original relative paths, the relative path in the alternates file
will point to the pool also in the snapshot without requiring us to rewrite
the path.