log: Move PayloadBytes tests into grpcstats package
Our log package hosts a test that verifies whether we can correctly log
the bytes of the request and response payloads. This logic is not part
of the log package though, but is instead implemented by the grpcstats
package.

Let's move the test to be hosted in the correct package such that the
system under test and the test itself are close together. This also
addresses an upcoming import cycle in our tests between the grpcstats
and the log package.