hook: Make separate contexts in tests more explicit
One of our tests for the pack-objects hook exercises what happens when
the context of the original Goroutine that created a cache entry gets
cancelled. The logic to do this is really hard to follow though as there
are multiple contexts, streams and waiters, and neither of them are well
encapsulated.

Refactor the code to clearly separate the logic of both calls by putting
all related code into separate Goroutines. This allows the reader to
understand much more easily what setup belongs to which of both callers.