hook: Tighten error check for pack-objects hook with cancelled context
We don't verify the error we receive in one of our tests that checks for
cancelled contexts, but only check for the error code. This isn't all
that helpful though in case the test would fail due to different errors
as the observer has no way to see what was actually different.

Refactor the code to instead explicitly verify the error message. Note
that we also drop the expectation that we might receive different errors
on macOS. The underlying race actually triggers for other Unix systems
as well and will be addressed in a subsequent commit.