hook: Synchronize test logic better in context cancellation test
The setup for our context cancellation test for the pack-objects hook is
quite involved. But regardless of that, we still don't have full control
over the actual steps.

Let's synchronize the test logic much more tightly and annotate the
assumptions we do at each step. Hopefully, this should help us to fix
the flaky tests.