hook: Fix invalid setup in pack-objects context cancellation test
Even though there is a comment stating that we write a commit with a
largish blob, we actually don't. This issue can cause racy behaviour
because the whole response would fit into a single message, and thus it
can happen that git-pack-objects(1) exits successfully before we can
actually verify its behaviour in a second Goroutine.

Fix the test setup to again write a largish blob, where we use a factor
of the write buffer size that the underlying stream would be using. This
should hopefully fix the flaky behaviour we have seen in this test.