trace2: Don't round offset event times
Currently when adding a relative (t_rel) or absolute (t_abs) time events
to a span's starting time, we inadvertently round it to the closest full
second. As most trace events are only a few milliseconds long, this
results in the complete loss of precision for that span.

Fix time addition to use fractional seconds.

Changelog: fixed