log: Provide logging functions that have a context available
One of the changes that the `slog` package will bring with it is the
ability to pass a context to the logging functions via `DebugContext()`
and similar function calls. The intent here is that the log handler can
extract information from that context and append it as structured data
to the logged record.

This change causes some ripples in the ecosystem. Most importantly, the
go-grpc-middleware dependency removed the `ctxlogrus` package completely
in version 2. The ability to inject logging fields into the context is
thus gone, and we are forced to adapt.

Given that we're already moving into the direction of adopting the slog
logger it makes sense to provide the same new calling convention where
you can inject the context into the logging calls. This commit thus
implements said new interface. For now, we still rely on the `ctxlogrus`
package to carry injected logging fields. At a later point though we can
replace this with our own holder for log fields once we can get rid of
the `ctxlogrus` package completely.