command: Recognize nested exit status
The `command.ExitStatus()` function is responsible for extracting an
exit code from an `exec.ExitError`. The way this is done is extremely
fragile though as we try to directly cast the passed-in error to the
target error type. It's quite likely though that the error may be
wrapped somewhere in the callchain, which would break the functionality.

Refactor the function to instead use `errors.As()` such that it knows to
peel the error chain correctly.