ssh: Rename `monitorStdinCommand()` to `runUploadCommand()`
With the preceding refactorings of `monitorStdinCommand()` its current
name is not quite accurate anymore: it does not only set up monitoring
anymore, but also runs the command and handles various bits and pieces
of its lifecycle.

Let's rename the function to `runUploadCommand()`. At a later point we
can also move this functionality into a shared place such that it is
reusable for the smarthttp service. While at it, document what the
functions intent is.