Flaky tests: Fix TestAdaptiveCalculator flaky tests
In TestAdaptiveCalculator, we set the calibration duration to 10
milliseconds. This test setup uses a manual ticker. This calibration
duration is irrelevant to the actual calibration cycle of the
calculation. That value was supposed to be a placeholder.

Unfortunately, the calculator uses this value to determine a timeout
when polling events from the watchers. Although event polling is
supposed to be instant, it might not be the case for slow machines.
Thus, on CIs, even polling might return deadline exceeded
unintentionally. This commit fixes this flake by setting the calibration
duration to an unrealistically high value.

For https://gitlab.com/gitlab-org/gitaly/-/issues/5467.