global: Inject logger into service servers
Many callsites extract the logger via the passed-in context by using the
`ctxlogrus` package. The ability to add loggers to the context via this
package is going away upstream though in favor of a new set of logging
functions that accept a context as input parameter themselves.

We're thus going to introduce an equivalent interface, but that will
require our service servers to have a logger available to them. Prepare
for this change by injecting the logger into both Gitaly and Praefect
servers.