Rename the closed channel on partition
Partition's closed channel indicates its TransactionManager has stopped
running. This is used to determine when it is safe to start another
TransactionManager for the partition as there should only ever be one
running. The 'closed' channel is confusingly named as it can be closed
before the 'closing' channel of the partition. The 'closing' channel
indicates the partition has started to shutdown and new transactions
should not be begun on it.

Rename the 'closed' channel to 'transactionManagerClosed' so it won't
be confused with the partition's closed state.