Assert all repositories as seen by a transaction
This commit expands the assertions we do in TransactionManager's
tests to cover transactions view of all repositories. TransactionManager
still only supports a single repository, so all transactions that one
repository. With this change, we'll be in position to assert a transaction's
view of multiple repositories as we add support for it.