Instrument start up time
Gitaly is starting exceedingly slow in some cases. We don't have
enough logs where the slow down happens to say which step in the
initialization process is the culprit. This commit outpus log
messages with the duration it took to execute a given initialization
step so we'll have a better idea what is taking so long.