Write quarantined objects into transaction's target repository
TransactionManager's tests are writing quarantined objects always
into the same repository. This is fine so far as the TransactionManager
still works on just a single repository. The transaction now contains
the target repository's relative path. To prepare for operating and
testing on multiple repositories, change the tests to write the
quarantined objects into the transaction's target repository.