log: Remove Print and related functions from `Logger` interface
The function family `Print{,f,ln}()` is merely a wrapper of the
respective functions that log a message at info level. Despite being
duplicate, it's also not obvious what they are doing and not supported
by the `slog` package.

Remove these functions from the `Logger` interface and replace all their
usages with the respective info-level functions. As the Yamux library
expects a logger implementing these functions this commit implements a
wrapper that forwards the respective functions.