log: Remove Warning and related functions from `Logger` interface
The `logrus.Logger` interface has both `Warn()` and `Warning()` and
related functions, and we have adopted these for our own `Logger`
interface. Let's remove the latter family of functions so that we only
keep the `Warn()` family of functions, which is also the interface
exposed by the `slog` package.

Note that the `badger` dependency requires the `Warning()`-style
functions. To address this, we again create a simple wrapper that
implements this specific interface.