log: Remove `Debugln()` and related functions from `Logger` interface
Drop `Debugln()` and related functions. The only difference to the
`Debug()` family of functions is that they use `fmt.Sprintln()` to
format the log message, which guarantees that the arguments will be
separated by a space. This difference is non-obvious and also not
supported by the `slog` package.

The only user of this function is the Yamux log wrapper, which we adapt
accordingly.