Allow snapshotting multiple repositories
Transactions currently snapshot the target repository and the alternate
repository if it has one. This allows transactions to operate on
repositories that already have an alternate but doesn't yet make it
possible add new alternate links. To do so, we'd need to be able to
start a transaction covering both the target repository, and the
alternate repository it is about to be linked to.

This commit makes it possible to start a transaction covering multiple
repositories. The transaction has a single target repository where
changes are recorded. In addition, multiple other repositories can be
included in the snapshot. Changes into these repositories are not
recorded or committed as we don't have a need to do so yet.

With the pool and the new member in the snapshot, the RPC handler
managing alternate links can create the link and verify it in its own
snapshot, and stage the change for committing. This will be implemented
later.