Assert repository's alternate
Repositories may have an alternate configured in 'objects/info/alternates'.
The path there is a relative path from '<repo>/objects' to
'<alternate-repo>/objects'. As we're about to support setting an alternate
in a transaction, this commit extends the repository state assertions in
TransactionManager tests to also assert the contents of the alternate file.