Support updating repository's alternate in a transaction
Repositories are linked to object pools via the 'objects/info/alternates'
file. Currently there's no way to stage an update to the alternates file
in a transaction. This commit adds support for staging an update which
allows for either setting an alternate or removing an existing. Basic
consistency checks are done to prevent relinking repositories that already
have an alternate and unlinking repositories that don't have one.