gitaly/config: Add negotiation timeouts to config
Adds a new `timeout` section to the TOML configuration file. This allows
users to customise the timeout values used for the negotiation phase of
the `git-upload-pack` and `git-upload-archive` operations, which may
time out for larger repositories.

The current default values of 10 minutes for git-upload-pack and 1
minute for git-upload-archive are preserved.