limiter: Integrate adaptive limit to pack-objects limiter
This commit integrates the adaptive limit to pack-objects limiter of the
Gitaly server. It introduces some new configurations for pack-objects
limiting. When Gitaly server starts, it creates an adaptive calculator.
The calculator adjusts the adaptive limit which is already integrated
into the concurrency limit in prior commits.

To enable the adaptive limiter, the operator needs to set
`pack_objects_limiting.adaptive` config to true and corresponding limits
located at the same location.