Setup TransactionManager's state and staging directories in storage
TransactionManager's tests are currently not setting up the staging
and state directories as they'd be configured in production. They're
generally set up by the PartitionManager which is not used in
TransactionManager's tests. This commit configures the tests to set
up both directories withing the storage directory. This is necessary
as we're about to do some relative path mangling which will verify
that relative paths don't point outside of the storage.