limiter: Use initial limit for static semaphore
Recently, the limiter's static semaphore uses the current value of limit
setting. When the adaptive calculator is not connected to the setting,
the current value is always the same as the initial limit. When it does,
it's not the case if the adaptive is enabled. This commit lets the
limiter uses the initial limit for static semaphore. It acts as an
escape hatch so that we can disable the adaptiveness and fallback to use
the static limiter when something goes wrong.