Change admission to a boolean
Transaction's admission field tracks whether it has been dequeued by
the TransactionManager for processing. If so, the TransactionManager
is now responsible for the clean up of the transaction. This was a
channel as we previously had to wait for a transaction to be dequeued
in tests to ensure a repository deletion has been processed where we
expect it to. As deletions are no longer done async in tests, change
the admission field back to a simple boolean.