git: Inject context into per-command options to allow feature for flags
We're about to introduce a feature flag that adds a Git configuration
key to just a single command. We don't really have a nice way to do this
right now though because the per-command options we have don't accept a
context.

Refactor the code so that we have a context available and can make use
of feature flags.