middleware: Rename metadatahandler to requestinfohandler
We have two different mechanisms to inject information about the current
gRPC call into the context:

    - The metadatahandler injects data about the metadata, e.g. the peer
      information and whether a call is an accessor or a mutator.

    - The field extractor that gets set up via go-grpc-middleware's tags
      interceptor and that injects data  about the request.

The tagging mechanism used by the field extractor is going away in the
go-grpc-middleware dependency though without replacement. Furthermore,
both functionalities are related to each other and can thus easily share
the same infrastructure. We are thus about to merge them.

As a first step, let's rename the metadatahandler to requestinfohandler
such that it better reflects the new reponsibility when it will start to
handle both gRPC call metadata and request information.