fieldextractors: Stop injecting namespace-service related tags
The fieldextractors take the initial gRPC request, extract important
information like the repository scope or the object pool, and then
inject this information into our tags. Like this, the information will
be picked up by our logging infrastructure when used with the ctxlogrus
package.

Part of what we report is related to namespace requests. As we are about
to merge the fieldextractors into our requestinfohandler, we'd have to
also carry over these kinds of requests. The namespace service is not in
use nowadays though anymore and will soon be removed, so porting over
this logic would be a waste of time.

Let's remove the logic and be done with it.