requesthandler: Inject information into the context
Throughout our codebase we have a bunch of callsites that rely on the
tags injected by our requestinfohandler package in order to generically
derive information about the current RPC call's context. This mechanism
is fragile as it relies on string keys to extract the information, where
it's not clear where the information actually comes from. Second, the
tagging mechanism is going away in go-grpc-middleware v2.

Start injecting the request information structure into the context. Like
this, we can adapt callsites to use this structure instead of tags.