sentryhandler: Allow deterministic testing of `generateSentryEvent()`
In order to calculate the duration that a gRPC call took we pass in the
start time to `generateSentryEvent()` from which we derive the duration.
This makes it hard to test this function deterministically though as the
duration is derived from the current time via `time.Since()`.

Refactor the code to instead pass in a duration such that we can test
this function deterministically.