sentryhandler: More rigidly test generated events
The way we test generated events is quite loose because we only assert
prats of the generated event's members. Next to being a bit sloppey, it
also has the effect that the test setup is harder to understand as there
are a bunch of variables per testcase that describe our expectations.

Refactor the test to instead assert the full expected event.