sentryhandler: Fix test for `MarkToSkip()` not doing anything
The `MarkToSkip()` function marks an RPC call in such a way that the
sentryhandler will not report the error to Sentry anymore. And while we
do have a test for this case, the test doesn't do anything because the
testcase doesn't set up an error. Consequentially, even if we didn't
call `MarkToSkip()` we wouldn't see any reported error.

Fix the test by returning an error for that testcase. While at it, add a
second test that has the exact same setup without calling `MarkToSkip()`
such that we can demonstrate that the call does indeed have an effect.