sentryhandler: Allow overriding function that reports events
We don't have any tests that verify that the sentryhandler interceptors
work as expected. With the current layout this is also quite impossible
to do, because the event that is to be reported will be passed to the
`sentry.CaptureEvent()` function without giving us a way to intercept
the captured event.

Refactor the code such that it becomes possible to replace the function
that reports events. Like that, we can intercept the capture function
and thus exercise the code more readily.