sentryhandler: Refactor skipping of logs to not use tags
In v2 of grpc-go-middleware, the tags functionality that allowed
middleware to add arbitrary information to the request context has been
removed without replacement. In our sentryhandler though we use this
mechanism in order to mark RPCs that should not be logged via Sentry.

Refactor the code to use our own context key that we inject into the
context via which we track whether logging should be skipped. This is
another step towards the upgrade to grpc-go-middleware v2.