tempdir: don't launch a go routine when creating a temporary directory
temdir package is launchig a goroutine every time a temporary directory
is created. The purpose of the goroutine is to clean up the temporary
directory on context cancellation. This doesn't seem very necessary and
has the following downsides:

- In our tests, we look for goroutines that are still running after
tests. This deletion goroutine is not synchronized with this and can
thus cause flskes.

- If context is canceled, the directory is deleted. If there is still
code operating on the directory, that code can fail with unexpected
errors rather than returning a context canceled error.

So instead defer the deletion of the directory when it is created.

Signed-off-by: Siddharth Asthana <siddharthasthana31@gmail.com>